+++ src/main/java/net/silentchaos512/gems/block/supercharger/TileSupercharger.java
@@ -20,33 +20,35 @@ package net.silentchaos512.gems.block.supercharger;
 
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.enchantment.EnchantmentHelper;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.inventory.IInventory;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.ITickable;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.TextComponentTranslation;
 import net.silentchaos512.gems.SilentGems;
 import net.silentchaos512.gems.compat.gear.SGearProxy;
 import net.silentchaos512.gems.init.ModEnchantments;
 import net.silentchaos512.gems.init.ModTags;
 import net.silentchaos512.gems.init.ModTileEntities;
-import net.silentchaos512.lib.util.MathUtils;
+import net.silentchaos512.lib.tile.SyncVariable;
+import net.silentchaos512.lib.tile.TileSidedInventorySL;
+import net.silentchaos512.lib.util.TimeUtils;
+import net.silentchaos512.utils.MathUtils;
 
 import javax.annotation.Nullable;
 
-public class TileSupercharger extends TileEntity implements ITickable, IInventory /* TileSidedInventorySL implements ITickable, IChaosAccepter */ {
+public class TileSupercharger extends TileSidedInventorySL implements ITickable {
     static final int MAX_CHAOS_STORED = 1_000_000;
     private static final int INVENTORY_SIZE = 3;
-    private static final int UPDATE_FREQUENCY = 15 * 20; // TimeHelper.ticksFromSeconds(15);
+    private static final int UPDATE_FREQUENCY =  TimeUtils.ticksFromSeconds(15);
 
-    //    @SyncVariable(name = "Energy")
+    @SyncVariable(name = "Energy")
     private int chaosStored;
-    //    @SyncVariable(name = "Progress")
+    @SyncVariable(name = "Progress")
     private int progress;
-    //    @SyncVariable(name = "StructureLevel")
+    @SyncVariable(name = "StructureLevel")
     private int structureLevel;
     private int updateTimer = 0;
 
@@ -65,50 +67,52 @@ public class TileSupercharger extends TileEntity implements ITickable, IInventor
                 SilentGems.LOGGER.info("Supercharger at {}: structure level updated to {}",
                         this.pos, this.structureLevel);
             }
-//            sendUpdate();
+            sendUpdate();
             updateTimer = 0;
         }
 
         ItemStack input = getInputItem();
         ItemStack catalyst = getCatalystItem();
 
+        // temp hack to make it work
+        receiveCharge(10000, false);
+
         if (!input.isEmpty() && !catalyst.isEmpty())
             handleCharging(input, catalyst);
         else progress = 0;
-
     }
 
     private void handleCharging(ItemStack input, ItemStack catalyst) {
-//        int chargeTier = ModItems.chargingAgent.getTier(catalyst);
-//        if (chargeTier > 0 && chargeTier <= this.structureLevel) {
-//            int totalCost = getChaosCostForCharging();
-//            int chaosDrained = MathUtils.min(chaosStored,
-//                    getChaosDrainPerTick(),
-//                    totalCost - progress);
-//
-//            if (wouldFitInOutputSlot(input, chargeTier)) {
-//                if (chaosDrained > 0) {
-//                    chaosStored -= chaosDrained;
-//                    progress += chaosDrained;
-//                }
-//                if (progress >= totalCost) {
-//                    if (getStackInSlot(2).isEmpty()) {
-//                        ItemStack output = input.copy();
-//                        output.setCount(1);
-//                        output.addEnchantment(ModEnchantments.supercharged, chargeTier);
-//                        setInventorySlotContents(2, output);
-//                    } else {
-//                        getStackInSlot(2).grow(1);
-//                    }
-//
-//                    progress = 0;
-//                    decrStackSize(0, 1);
-//                    decrStackSize(1, 1);
-//                }
-//
-//                if (chaosDrained > 0) sendUpdate();
-//            }
-//        }
+        int chargeTier = getChargingAgentTier(catalyst);
+        if (chargeTier > 0 && chargeTier <= this.structureLevel) {
+            int totalCost = getChaosCostForCharging();
+            int chaosDrained = MathUtils.min(chaosStored,
+                    getChaosDrainPerTick(),
+                    totalCost - progress);
+
+            if (wouldFitInOutputSlot(input, chargeTier)) {
+                if (chaosDrained > 0) {
+                    chaosStored -= chaosDrained;
+                    progress += chaosDrained;
+                }
+                if (progress >= totalCost) {
+                    if (getStackInSlot(2).isEmpty()) {
+                        ItemStack output = input.copy();
+                        output.setCount(1);
+                        output.addEnchantment(ModEnchantments.supercharged, chargeTier);
+                        setInventorySlotContents(2, output);
+                    } else {
+                        getStackInSlot(2).grow(1);
+                    }
+
+                    progress = 0;
+                    decrStackSize(0, 1);
+                    decrStackSize(1, 1);
+                }
+
+                if (chaosDrained > 0) sendUpdate();
+            }
+        }
     }
 
     private ItemStack getInputItem() {
@@ -117,9 +121,8 @@ public class TileSupercharger extends TileEntity implements ITickable, IInventor
     }
 
     private ItemStack getCatalystItem() {
-//        ItemStack stack = getStackInSlot(1);
-//        return stack.getItem() == ModItems.chargingAgent ? stack : ItemStack.EMPTY;
-        return ItemStack.EMPTY;
+        ItemStack stack = getStackInSlot(1);
+        return getChargingAgentTier(stack) > 0 ? stack : ItemStack.EMPTY;
     }
 
     private boolean checkStructureLevel() {
@@ -157,6 +160,17 @@ public class TileSupercharger extends TileEntity implements ITickable, IInventor
         return 0;
     }
 
+    private static int getChargingAgentTier(ItemStack stack) {
+        Item item = stack.getItem();
+        if (item.isIn(ModTags.Items.CHARGING_AGENT_TIER3))
+            return 3;
+        if (item.isIn(ModTags.Items.CHARGING_AGENT_TIER2))
+            return 2;
+        if (item.isIn(ModTags.Items.CHARGING_AGENT_TIER1))
+            return 1;
+        return 0;
+    }
+
     private boolean wouldFitInOutputSlot(ItemStack input, int chargeTier) {
         ItemStack output = getStackInSlot(2);
         if (output.isEmpty()) return true;
@@ -172,22 +186,20 @@ public class TileSupercharger extends TileEntity implements ITickable, IInventor
 
     @Deprecated
     private static int getChaosCostForCharging(ItemStack gearPart, ItemStack catalyst) {
-//        if (gearPart.isEmpty() || catalyst.isEmpty()) return -1;
-//        final int tier = SGearProxy.getPartTier(gearPart);
-//        final int chargeTier = ModItems.chargingAgent.getTier(catalyst);
-//        return tier >= 0 && chargeTier > 0 ? chargeTier * (chargeTier + 1) * (int) Math.pow(10, tier + 1) : -1;
-        return 0;
+        if (gearPart.isEmpty() || catalyst.isEmpty()) return -1;
+        final int tier = SGearProxy.getPartTier(gearPart);
+        final int chargeTier = getChargingAgentTier(catalyst);
+        return tier >= 0 && chargeTier > 0 ? chargeTier * (chargeTier + 1) * (int) Math.pow(10, tier + 1) : -1;
     }
 
     @Deprecated
     int getChaosDrainPerTick() {
-//        final int partTier = SGearProxy.getPartTier(this.getInputItem());
-//        final int chargeTier = ModItems.chargingAgent.getTier(this.getCatalystItem());
-//        if (partTier < 0 || chargeTier <= 0) return 0;
-//        int rate = getChaosDrainPerTick(partTier, chargeTier);
-//        // Limit the rate so low tier materials don't supercharge instantly
-//        return Math.min(rate, getChaosCostForCharging() / 20);
-        return 0;
+        final int partTier = SGearProxy.getPartTier(this.getInputItem());
+        final int chargeTier = getChargingAgentTier(this.getCatalystItem());
+        if (partTier < 0 || chargeTier <= 0) return 0;
+        int rate = getChaosDrainPerTick(partTier, chargeTier);
+        // Limit the rate so low tier materials don't supercharge instantly
+        return Math.min(rate, getChaosCostForCharging() / 20);
     }
 
     @Deprecated
@@ -198,32 +210,33 @@ public class TileSupercharger extends TileEntity implements ITickable, IInventor
 
     @Override
     public boolean isItemValidForSlot(int index, ItemStack stack) {
-//        if (index == 0) return SGearProxy.isMainPart(stack);
-//        if (index == 1) return stack.getItem() == ModItems.chargingAgent;
+        if (index == 0) return SGearProxy.isMainPart(stack);
+        if (index == 1) return stack.getItem().isIn(ModTags.Items.CHARGING_AGENTS);
         return false;
     }
 
 //    @Override
-//    public int receiveCharge(int maxReceive, boolean simulate) {
-//        int received = Math.min(MAX_CHAOS_STORED - chaosStored, maxReceive);
-//        if (!simulate) {
-//            chaosStored += received;
-//            if (received != 0) {
-//                sendUpdate();
-//            }
-//        }
-//        return received;
-//    }
-//
+    public int receiveCharge(int maxReceive, boolean simulate) {
+        int received = Math.min(MAX_CHAOS_STORED - chaosStored, maxReceive);
+        if (!simulate) {
+            chaosStored += received;
+            if (received != 0) {
+                sendUpdate();
+            }
+        }
+        return received;
+    }
+
 //    @Override
-//    public int getCharge() {
+    public int getCharge() {
 //        return chaosStored;
-//    }
-//
+        return MAX_CHAOS_STORED;
+    }
+
 //    @Override
-//    public int getMaxCharge() {
-//        return MAX_CHAOS_STORED;
-//    }
+    public int getMaxCharge() {
+        return MAX_CHAOS_STORED;
+    }
 
     @Override
     public int getSizeInventory() {
@@ -232,47 +245,8 @@ public class TileSupercharger extends TileEntity implements ITickable, IInventor
 
     @Override
     public boolean isEmpty() {
+        // TODO
         return false;
-    }
-
-    @Override
-    public ItemStack getStackInSlot(int index) {
-        return null;
-    }
-
-    @Override
-    public ItemStack decrStackSize(int index, int count) {
-        return null;
-    }
-
-    @Override
-    public ItemStack removeStackFromSlot(int index) {
-        return null;
-    }
-
-    @Override
-    public void setInventorySlotContents(int index, ItemStack stack) {
-
-    }
-
-    @Override
-    public int getInventoryStackLimit() {
-        return 0;
-    }
-
-    @Override
-    public boolean isUsableByPlayer(EntityPlayer player) {
-        return false;
-    }
-
-    @Override
-    public void openInventory(EntityPlayer player) {
-
-    }
-
-    @Override
-    public void closeInventory(EntityPlayer player) {
-
     }
 
     @Override
@@ -282,7 +256,6 @@ public class TileSupercharger extends TileEntity implements ITickable, IInventor
 
     @Override
     public void clear() {
-
     }
 
     @Override
@@ -300,7 +273,7 @@ public class TileSupercharger extends TileEntity implements ITickable, IInventor
 
     @Override
     public ITextComponent getName() {
-        return null;
+        return new TextComponentTranslation("container.silentgems.supercharger");
     }
 
     @Override